# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/login":{"post":{"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons":{"get":{"summary":"List Persons","operationId":"list_persons_persons_get","parameters":[{"name":"tipo","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tipo"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonaOut"},"title":"Response List Persons Persons Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Person","operationId":"create_person_persons_post","parameters":[{"name":"x-user-role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{id}":{"get":{"summary":"Get Person","operationId":"get_person_persons__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Person","operationId":"update_person_persons__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"x-user-role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Person","operationId":"delete_person_persons__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"x-user-role","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Role"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginIn":{"properties":{"username":{"type":"string","title":"Username"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}},"type":"object","required":["username"],"title":"LoginIn"},"PersonaIn":{"properties":{"nombres_persona":{"type":"string","maxLength":50,"title":"Nombres Persona"},"apellido_paternoPersona":{"anyOf":[{"type":"string","maxLength":30},{"type":"null"}],"title":"Apellido Paternopersona"},"apellido_maternoPer":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Apellido Maternoper"},"telefono_persona":{"anyOf":[{"type":"string","maxLength":15},{"type":"null"}],"title":"Telefono Persona"},"id_tipoPersona":{"type":"integer","title":"Id Tipopersona"},"ci_persona":{"type":"string","maxLength":10,"title":"Ci Persona"},"direccion_persona":{"type":"string","maxLength":100,"title":"Direccion Persona"}},"type":"object","required":["nombres_persona","id_tipoPersona","ci_persona","direccion_persona"],"title":"PersonaIn"},"PersonaOut":{"properties":{"nombres_persona":{"type":"string","maxLength":50,"title":"Nombres Persona"},"apellido_paternoPersona":{"anyOf":[{"type":"string","maxLength":30},{"type":"null"}],"title":"Apellido Paternopersona"},"apellido_maternoPer":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Apellido Maternoper"},"telefono_persona":{"anyOf":[{"type":"string","maxLength":15},{"type":"null"}],"title":"Telefono Persona"},"id_tipoPersona":{"type":"integer","title":"Id Tipopersona"},"ci_persona":{"type":"string","maxLength":10,"title":"Ci Persona"},"direccion_persona":{"type":"string","maxLength":100,"title":"Direccion Persona"},"id_persona":{"type":"integer","title":"Id Persona"}},"type":"object","required":["nombres_persona","id_tipoPersona","ci_persona","direccion_persona","id_persona"],"title":"PersonaOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}